backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pydantic==2.10.4
pydantic-settings==2.7.1
celery[redis]==5.4.0
redis==5.2.1
pymilvus==2.4.9
minio==7.2.12
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
PyMuPDF==1.25.1
pdfplumber==0.11.4
crewai==0.86.0
httpx==0.28.1
python-multipart==0.0.20
python-dotenv==1.0.1
websockets==14.1
marshmallow>=3.18.0,<3.25.0
environs>=9.5.0,<12.0.0
psycopg2-binary==2.9.10
